requirements.txt
psycopg2==2.9.3
Django==4.0.5
graphene-django
pylint==2.14.0
black==22.3.0
pydantic==1.9.1
django-treebeard==4.5.1
django-extensions==3.1.5
